move clientManager related to core module && add RocketmqService
