+++ eventmesh-dashboard-core/src/main/java/org/apache/eventmesh/dashboard/core/store/RocketmqTopicCore.java
@@ -18,9 +18,9 @@
 package org.apache.eventmesh.dashboard.core.store;
 
 
-import org.apache.eventmesh.dashboard.common.properties.RocketmqProperties;
 import org.apache.eventmesh.dashboard.common.dto.TopicProperties;
-import org.apache.eventmesh.dashboard.common.util.RocketmqUtils;
+import org.apache.eventmesh.dashboard.common.properties.RocketmqProperties;
+import org.apache.eventmesh.dashboard.service.mq.RocketmqTopicService;
 import org.apache.eventmesh.dashboard.service.store.TopicCore;
 
 import org.apache.rocketmq.common.TopicFilterType;
@@ -28,6 +28,7 @@ import org.apache.rocketmq.common.constant.PermName;
 
 import java.util.List;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import lombok.extern.slf4j.Slf4j;
@@ -35,6 +36,9 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 @Service
 public class RocketmqTopicCore implements TopicCore {
+
+    @Autowired
+    private RocketmqTopicService rocketmqTopicService;
 
     private final RocketmqProperties rocketmqProperties;
 
@@ -44,12 +48,12 @@ public class RocketmqTopicCore implements TopicCore {
 
     @Override
     public List<TopicProperties> getTopics() {
-        return RocketmqUtils.getTopics(rocketmqProperties.getNamesrvAddr(), rocketmqProperties.getRequestTimeoutMillis());
+        return rocketmqTopicService.getTopics(rocketmqProperties.getNamesrvAddr(), rocketmqProperties.getRequestTimeoutMillis());
     }
 
     @Override
     public void createTopic(String topicName) {
-        RocketmqUtils.createTopic(topicName, TopicFilterType.SINGLE_TAG.name(),
+        rocketmqTopicService.createTopic(topicName, TopicFilterType.SINGLE_TAG.name(),
             PermName.PERM_READ | PermName.PERM_WRITE, rocketmqProperties.getNamesrvAddr(),
             rocketmqProperties.getReadQueueNums(), rocketmqProperties.getWriteQueueNums(),
             rocketmqProperties.getRequestTimeoutMillis());
@@ -57,7 +61,7 @@ public class RocketmqTopicCore implements TopicCore {
 
     @Override
     public void deleteTopic(String topicName) {
-        RocketmqUtils.deleteTopic(topicName, rocketmqProperties.getNamesrvAddr(),
+        rocketmqTopicService.deleteTopic(topicName, rocketmqProperties.getNamesrvAddr(),
             rocketmqProperties.getRequestTimeoutMillis());
     }
 }
